Added support for creating users and vdc in a given organization.
This allows for creating a complete environment on vcd i.e. Create a new org, create a user in the new org,
create a vdc in the new org. Previously this was not possible as the user and vdc modules would create the
resources in the org in which the logged in user resides.

Signed-off-by: Joost Coelingh <joost.coelingh@eu.equinix.com>